+++ examples/go-consumer/consumer_test.go
@@ -1,0 +1,9 @@
+package consumer
+
+import "testing"
+
+func TestName(t *testing.T) {
+    if Name() != "go-consumer" {
+        t.Fatalf("unexpected name: %s", Name())
+    }
+}